+++ tests/dist.R
@@ -21,7 +21,7 @@ crs = 4326
 
 st_distance(x, y)
 
-d.sf = st_distance(x, y, geosphere::distMeeus)
+d.sf = st_distance(x, y, dist_fun = geosphere::distMeeus)
 d.sp = spDists(as(x, "Spatial"), as(y, "Spatial"))
 units(d.sp) = make_unit("km")
 d.sf - d.sp